Refactor Footer component link in Footer.jsx
The code changes in Footer.jsx update the link in the footer component to point to the AWS website. This change reflects the updated partnership and improves the accuracy of the information displayed to users. By linking to the AWS website, users can access more information about the technology powering the application. This refactor enhances the professionalism and credibility of the application.

Note: This is a suggested commit message based on the provided code changes.