Refactor NavigationBar component to remove unused CSS styles
The code changes in NavigationBar.css remove the 'text-decoration: none' style from the NavigationBar component. This style was not being used and can be safely removed. This refactor improves the code cleanliness and maintainability of the component.

Note: This is a suggested commit message based on the provided code changes.